+++ HISTORY.rst
@@ -1,6 +1,8 @@
 4.4.13 (XXXX-XX-XX)
 ===================
 - Update PIP from version 23.3.2 to 24.0.
+- Fix source class and JavaScript ``MayanImage`` class ``.initialize()``
+  method name.
 
 4.4.12 (2024-02-03)
 ===================